refactor: styling and cleanup (#18)
