+++ app.vue
@@ -1,7 +1,7 @@
 <template>
+	<vite-pwa-manifest />
+	<nuxt-loading-indicator />
 	<nuxt-layout name="main">
-		<vite-pwa-manifest />
-		<nuxt-loading-indicator />
 		<nuxt-page />
 	</nuxt-layout>
 </template>
